requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
fastapi>=0.110.0
uvicorn[standard]>=0.23.0
python-jose[cryptography]>=3.3.0
httpx>=0.27.0
